% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateDeltaMSSettings.R
\name{LocateDeltaMSSettings}
\alias{LocateDeltaMSSettings}
\title{Locate the shiny bookmarksfolder that contains saved parameter settings}
\usage{
LocateDeltaMSSettings(number = FALSE, names = FALSE)
}
\arguments{
\item{number}{Show number of saved settings (boolean, default is FALSE)}

\item{names}{Show names of saved settings (boolean, default is FALSE)}
}
\description{
Locate the shiny bookmarksfolder that contains saved parameter settings
}
\details{
\code{LocateDeltaMSSettings} optionally returns the number and the names of the present settings. If there are files of a certain setting missing, a message will be returned.
}
